refactor: ルーム分析文キャッシュをMySQLへ統合し /oc を1クエリ化、SQLiteを全廃
ルーム個別ページの分析文(narrative)キャッシュを SQLite から MySQL(ocgraph_ocreview*)へ移行。
getOpenChatByIdWithTag の oc_page_cache JOIN で open_chat と一緒に1クエリ取得するようにし、
/oc 表示時の SQLite 読み(と WAL/wal-index 競合による間欠ストール回避策)を丸ごと不要にした。

- MySQLテーブル oc_page_cache を ja/tw/th の ocreview schema に追加(narrative_data mediumtext)
- getOpenChatByIdWithTag に LEFT JOIN oc_page_cache + narrative_data を追加
- Controller は $oc['narrative_data'] から narrative を組み立て(SQLite read 廃止)
- OcPageCacheRepositoryInterface を新設し MySQL 実装を DI バインド。読みは JOIN に寄せ
  Repository は書き込み専用に。旧 SQLite 実装(SQLiteOcPageCache・repo・schema)を削除
- FileStorage設定・prod-sync同期対象・setupスクリプトから oc_page_cache(SQLite)を一掃
- 「分析文/関連ルーム」誤記を「分析文」に修正(関連ルームはこのキャッシュに含まれない)
- Repositoryは必ずInterfaceとセットで作る方針を CLAUDE.md に追記

Committed from: user-B550M-Pro4:~/repos/Open-Chat-Graph
