+++ batch/sh/prod-sync/lib/config.sh
@@ -95,7 +95,6 @@ SQLITE_DIRS=(
     ranking_position
     statistics_ohlc
     ranking_position_ohlc
-    oc_page_cache
 )
 SQLITE_SQLAPI_DIR="ocgraph_sqlapi"
