+++ setup/local-setup.default.sh
@@ -145,7 +145,6 @@ docker compose exec -T -u root app bash -c '
     rm -f storage/*/SQLite/statistics_ohlc/*.db*
     rm -f storage/*/SQLite/ranking_position/*.db*
     rm -f storage/*/SQLite/ranking_position_ohlc/*.db*
-    rm -f storage/*/SQLite/oc_page_cache/*.db*
     rm -f storage/*/SQLite/ocgraph_sqlapi/*.db*
     rm -f storage/*.log
     rm -f storage/*/*/*.log
@@ -263,9 +262,6 @@ for lang in ja tw th; do
 
     # statistics_ohlc.db を生成
     cat setup/schema/sqlite/statistics_ohlc.sql | docker compose exec -T -u www-data app sqlite3 "storage/${lang}/SQLite/statistics_ohlc/statistics_ohlc.db"
-
-    # oc_page_cache.db を生成（ルーム個別ページの分析文/関連ルーム事前計算キャッシュ）
-    cat setup/schema/sqlite/oc_page_cache.sql | docker compose exec -T -u www-data app sqlite3 "storage/${lang}/SQLite/oc_page_cache/oc_page_cache.db"
 
     # sqlapi.db を生成（jaのみ）
     if [ "$lang" == "ja" ]; then
